# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ILLICO DELIVERY API",
    "version": "1.0.0",
    "description": "Plateforme de livraison express & points de collecte — Cotonou & Abomey-Calavi\n\n## 🎯 Fonctionnalités\n- Authentification multi-rôles (Admin, Client, Livreur, Point ILLICO)\n- Gestion des livraisons en temps réel\n- Points de collecte avec stockage temporaire\n- Calcul tarifaire dynamique\n- Paiements & commissions\n- Notifications push\n\n## 🔐 Authentification\nToutes les routes protégées nécessitent un token JWT dans le header :\n```\nAuthorization: Bearer <votre_token_jwt>\n```\n\n## 🌍 Base URL\n- Development: `http://localhost:3000/api`\n- Production: `https://illico-delivery.vercel.app/api`"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development (Local)"
    },
    {
      "url": "https://illico-delivery.vercel.app/api",
      "description": "Production (Vercel)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenu après connexion. Format: `Bearer <token>`"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "description": "Utilisateur (4 rôles possibles: Admin, Client, Livreur, PointIllico)",
        "properties": {
          "_id": { "type": "string", "example": "65f2a1b2c3d4e5f6a7b8c9d0" },
          "nom": { "type": "string", "example": "Koffi Mensah" },
          "telephone": { "type": "string", "example": "+22997123456" },
          "email": { "type": "string", "example": "admin@illicodelivery.bj", "nullable": true },
          "role": { "type": "string", "enum": ["Admin", "Client", "Livreur", "PointIllico"], "example": "Client" },
          "adresse": { "type": "string", "example": "Cotonou, Haie Vive, Rue 123", "nullable": true },
          "typeClient": { "type": "string", "enum": ["standard", "professionnel"], "example": "professionnel", "nullable": true },
          "soldeIllico": { "type": "number", "example": 15000, "nullable": true },
          "forfaitActif": { "type": "string", "nullable": true, "example": "65f2a1b2c3d4e5f6a7b8c9d1" },
          "vehicule": { "type": "object", "nullable": true, "example": { "_id": "65f2a1b2c3d4e5f6a7b8c9d2", "type": "moto", "tarifBase": 500 } },
          "statut": { "type": "string", "enum": ["en_ligne", "hors_ligne", "en_mission"], "example": "en_ligne", "nullable": true },
          "scoreNote": { "type": "number", "minimum": 0, "maximum": 5, "example": 4.8, "nullable": true },
          "cashCollecte": { "type": "number", "example": 25000, "nullable": true },
          "cashReverse": { "type": "number", "example": 10000, "nullable": true },
          "valide": { "type": "boolean", "example": true, "nullable": true },
          "actif": { "type": "boolean", "example": true, "nullable": true },
          "commissionTotal": { "type": "number", "example": 5000, "nullable": true },
          "location": {
            "type": "object",
            "properties": {
              "type": { "type": "string", "example": "Point" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [2.4394, 6.3654] }
            }
          }
        }
      },
      "Livraison": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "65f2a1b2c3d4e5f6a7b8c9e0" },
          "client": { "$ref": "#/components/schemas/User" },
          "livreur": { "$ref": "#/components/schemas/User", "nullable": true },
          "pointDepart": {
            "type": "object",
            "properties": {
              "adresse": { "type": "string", "example": "Cotonou, Haie Vive" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [2.4394, 6.3654] }
            }
          },
          "pointArrivee": {
            "type": "object",
            "properties": {
              "adresse": { "type": "string", "example": "Abomey-Calavi, Godomey" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [2.3558, 6.4478] }
            }
          },
          "vehicule": { "type": "object", "properties": { "_id": { "type": "string" }, "type": { "type": "string" }, "tarifBase": { "type": "number" } } },
          "mode": { "type": "string", "enum": ["express", "point_illico"], "example": "express" },
          "pointIllico": { "$ref": "#/components/schemas/User", "nullable": true },
          "statut": { "type": "string", "enum": ["en_attente", "affecté", "arrivé_pickup", "colis_récupéré", "déposé_en_point", "livré", "échoué", "annulé"], "example": "en_attente" },
          "prixEstime": { "type": "number", "example": 2500 },
          "prixFinal": { "type": "number", "nullable": true, "example": 2500 },
          "modePaiement": { "type": "string", "enum": ["cash", "mobile_money", "credit_illico"], "example": "cash" },
          "otpLivraison": { "type": "string", "example": "123456", "writeOnly": true },
          "otpRetrait": { "type": "string", "example": "789012", "writeOnly": true },
          "otpValidé": { "type": "boolean", "example": false },
          "preuveLivraisonUrl": { "type": "string", "example": "https://res.cloudinary.com/.../proof.jpg", "nullable": true },
          "noteLivreur": { "type": "number", "minimum": 1, "maximum": 5, "nullable": true },
          "urgent": { "type": "boolean", "example": false },
          "nuit": { "type": "boolean", "example": false },
          "poids": { "type": "number", "example": 2.5 },
          "dateCreation": { "type": "string", "format": "date-time", "example": "2026-04-15T10:30:00Z" },
          "dateLivraison": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Identifiants invalides." },
          "errors": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["Email requis", "Mot de passe trop court"],
            "nullable": true
          }
        }
      },
...
      "LivraisonCreateRequest": {
        "type": "object",
        "required": ["pointDepart", "pointArrivee", "vehicule", "mode"],
        "properties": {
          "pointDepart": {
            "type": "object",
            "required": ["adresse", "coordinates"],
            "properties": {
              "adresse": { "type": "string", "example": "Cotonou, Haie Vive, Rue 123" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [2.4394, 6.3654], "description": "[longitude, latitude]" }
            }
          },
          "pointArrivee": {
            "type": "object",
            "required": ["adresse", "coordinates"],
            "properties": {
              "adresse": { "type": "string", "example": "Abomey-Calavi, Godomey, Carrefour" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [2.3558, 6.4478] }
            }
          },
          "vehicule": { "type": "string", "example": "65f2a1b2c3d4e5f6a7b8c9b0", "description": "ID du véhicule (Vehicule._id)" },
          "mode": { "type": "string", "enum": ["express", "point_illico"], "example": "express" },
          "pointIllico": { "type": "string", "nullable": true, "example": "65f2a1b2c3d4e5f6a7b8c9d3", "description": "ID du point ILLICO (si mode=point_illico)" },
          "urgent": { "type": "boolean", "example": false },
          "nuit": { "type": "boolean", "example": false },
          "poids": { "type": "number", "example": 2.5, "description": "Poids du colis en kg" },
          "modePaiement": { "type": "string", "enum": ["cash", "mobile_money", "credit_illico"], "example": "cash" }
        }
      },
...
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "ID MongoDB de la ressource",
        "example": "65f2a1b2c3d4e5f6a7b8c9d0"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Non authentifié",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "message": "Token invalide ou expiré." }
          }
        }
      },
...
      "NotFound": {
        "description": "Ressource non trouvée",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "message": "Livraison non trouvée." }
          }
        }
      },
      "BadRequest": {
        "description": "Requête invalide",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "message": "Champs requis manquants.", "errors": ["telephone requis"] }
          }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/livraisons": {
      "post": {
        "tags": ["📦 Livraisons"],
        "summary": "Créer une nouvelle livraison",
        "description": "Réservé aux clients et admins. Crée une livraison et déclenche le dispatch automatique vers un livreur disponible.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LivraisonCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Livraison créée",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Livraison créée avec succès.",
                  "data": {
                    "_id": "65f2a1b2c3d4e5f6a7b8c9e0",
                    "client": { "_id": "65f2a1b2c3d4e5f6a7b8c9d1", "nom": "Koffi Mensah", "telephone": "+22997123456" },
                    "statut": "en_attente",
                    "prixEstime": 2500,
                    "otpLivraison": "123***"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "get": {
        "tags": ["📦 Livraisons"],
        "summary": "Lister mes livraisons",
        "description": "Retourne les livraisons filtrées selon le rôle:\n- Client: ses propres livraisons\n- Livreur: ses missions\n- Admin: toutes les livraisons",
        "parameters": [
          { "name": "statut", "in": "query", "schema": { "type": "string", "enum": ["en_attente", "affecté", "livré", "annulé"] } },
          { "name": "mode", "in": "query", "schema": { "type": "string", "enum": ["express", "point_illico"] } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "example": 20 } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "example": 1 } }
        ],
        "responses": {
          "200": {
            "description": "Liste des livraisons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Livraison" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/livraisons/{id}": {
      "get": {
        "tags": ["📦 Livraisons"],
        "summary": "Détail d'une livraison",
        "parameters": [
          { "$ref": "#/components/parameters/id" }
        ],
        "responses": {
          "200": {
            "description": "Livraison récupérée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Livraison" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["📦 Livraisons"],
        "summary": "Annuler une livraison",
        "description": "Seul le client peut annuler sa livraison, et uniquement si le statut est 'en_attente'.",
        "parameters": [
          { "$ref": "#/components/parameters/id" }
        ],
        "responses": {
          "200": {
            "description": "Livraison annulée",
            "content": {
              "application/json": {
                "example": { "success": true, "message": "Livraison annulée avec succès." }
              }
            }
          },
          "400": { "description": "Impossible d'annuler (statut non éligible)" }
        }
      }
    },
...
  }
}